pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.17.0",
    "openai>=1.61.1",
    "python-dotenv>=1.0.1",
]
